requirements.txt
transformers
torch
nltk
pyfiglet
colorama
fpdf
fpdf2
pillow
pytesseract
tesseract